* tp/Texinfo/XS/main/debug.c (print_associate_info_debug), tp/Texinfo/XS/main/extra.c (lookup_extra_integer), tp/Texinfo/XS/main/manipulate_tree.c (increase_ref_counter) (get_copy_ref), tp/Texinfo/XS/main/tree_types.h (KEY_PAIR), tp/Texinfo/XS/main/utils.c (section_level), tp/Texinfo/XS/parsetexi/handle_commands.c (handle_other_command) (handle_line_command): use int for the integer field in KEY_PAIR values union. Update declarations and remove casts.

+++ tp/Texinfo/XS/main/debug.c
@@ -145,7 +145,7 @@ char *print_associate_info_debug (const ASSOCIATED_INFO *info)
           text_printf (&text, "deleted");
           break;
         case extra_integer:
-          text_printf (&text, "integer: %d", (int)k->integer);
+          text_printf (&text, "integer: %d", k->integer);
           break;
         case extra_string:
           text_printf (&text, "string: %s", k->string);
@@ -169,7 +169,7 @@ char *print_associate_info_debug (const ASSOCIATED_INFO *info)
               {
                 KEY_PAIR *k_integer = lookup_extra (f->contents.list[j], "integer");
                 if (k_integer)
-                  text_printf (&text, "%d|", (int)k_integer->integer);
+                  text_printf (&text, "%d|", k_integer->integer);
                 else
                   text_printf (&text, "%s|", f->contents.list[j]->text.text);
               }
